I upgraded Windows build to Visual Studio 2013 Desktop version, and wxWidgets 3.0

+++ trunk/src/OrientationModifier.cpp
@@ -24,10 +24,10 @@ void COrientationModifierParams::write_values_to_config()
 	HeeksConfig config;
 
 	// These values are in mm.
-	config.Write(_T("OrientationModifier_m_spacing"), m_spacing);
+	config.Write(_T("OrientationModifier_m_spacing"), (int)m_spacing);
 	config.Write(_T("OrientationModifier_number_of_rotations"), m_number_of_rotations);
 	config.Write(_T("OrientationModifier_sketch_rotates_text"), m_sketch_rotates_text);
-	config.Write(_T("OrientationModifier_justification"), m_justification);
+	config.Write(_T("OrientationModifier_justification"), (int)m_justification);
 }
 
 static void on_set_justification(int zero_based_choice, HeeksObj* object, bool from_undo_redo)
